+++ packages/ibm-products/src/components/FullPageError/FullPageError.stories.jsx
@@ -49,17 +49,10 @@ export default {
 
 const defaultProps = {
   kind: 'custom',
-  children: (
-    <>
-      <Link size="lg" href={'/'}>
-        – Forwarding Link 1
-      </Link>
-      <br />
-      <Link size="lg" href={'/'}>
-        – Forwarding Link 1
-      </Link>
-    </>
-  ),
+  title: '[Error title]',
+  label: 'Error ###',
+  description:
+    'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam.',
 };
 /**
  * TODO: Declare template(s) for one or more scenarios.
@@ -72,7 +65,19 @@ const Template = (args) => {
           <div className={`${storyClass}__breadcrumb-container`}>
             <Breadcrumbs className={`${storyClass}__breadcrumb`} />
           </div>
-          <FullPageError {...args} />
+          <FullPageError
+            title={`${args.title}`}
+            label={`${args.label}`}
+            description={`${args.description}`}
+          >
+            <Link size="lg" href={'/'}>
+              – Forwarding Link 1
+            </Link>
+            <br />
+            <Link size="lg" href={'/'}>
+              – Forwarding Link 1
+            </Link>
+          </FullPageError>
         </div>
       </UiShell>
     </div>
@@ -87,10 +92,6 @@ export const fullPageError = Template.bind({});
 fullPageError.storyName = 'Default';
 fullPageError.args = {
   ...defaultProps,
-  title: '[Error title]',
-  label: 'Error ###',
-  description:
-    'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam.',
 };
 export const fullPageError403 = Template.bind({});
 fullPageError403.storyName = '403';