+++ packages/ibm-products/src/components/InterstitialScreen/InterstitialScreen.test.js
@@ -6,18 +6,20 @@
  */
 
 import React from 'react';
-import { render, screen, act } from '@testing-library/react'; // https://testing-library.com/docs/react-testing-library/intro
+import { render, screen, act, waitFor } from '@testing-library/react'; // https://testing-library.com/docs/react-testing-library/intro
 
 import { pkg, carbon } from '../../settings';
 import uuidv4 from '../../global/js/utils/uuidv4';
 
 import { InterstitialScreen } from '.';
-import { InterstitialScreenView } from '..';
-import { InterstitialScreenViewModule } from '..';
 import userEvent from '@testing-library/user-event';
+import { InterstitialScreenViewModule } from './_story-assets/InterstitialScreenViewModule/InterstitialScreenViewModule';
+import { InterstitialScreenView } from './_story-assets/InterstitialScreenView/InterstitialScreenView';
 
 const blockClass = `${pkg.prefix}--interstitial-screen`;
 const componentName = InterstitialScreen.displayName;
+
+// cspell:words Terminé Partiel Actuel valide
 
 // values to use
 
@@ -26,29 +28,55 @@ const InterstitialScreenViewModuleTitle = `Title-${uuidv4()}`;
 const dataTestId = uuidv4();
 const { fn } = jest;
 const onClose = fn();
-const renderComponent = ({ ...rest } = {}) =>
-  render(
+const renderComponent = ({ ...rest } = {}) => {
+  const translations = {
+    'carbon.progress-step.complete': 'Terminé',
+    'carbon.progress-step.incomplete': 'Partiel',
+    'carbon.progress-step.current': 'Actuel',
+    'carbon.progress-step.invalid': 'Non valide',
+  };
+  const translateWithId = (messageId) => {
+    return translations[messageId];
+  };
+  return render(
     <InterstitialScreen
       isOpen={true}
       onClose={onClose}
       data-testid={dataTestId}
       {...{ ...rest }}
     >
-      <InterstitialScreenView stepTitle="Step 1">
-        <InterstitialScreenViewModule
-          title={InterstitialScreenViewModuleTitle}
-          description="Use case-specific content that explains the concept. Use case-specific content that explains the concept. Use case-specific content that explains the concept. Use case-specific content that explains the concept. Use case-specific content that explains the concept."
-        />
-      </InterstitialScreenView>
-
-      <InterstitialScreenView stepTitle="Step 2">
-        <InterstitialScreenViewModule
-          title="Use case-specific heading 2"
-          description="Use case-specific content that explains the concept. Use case-specific content that explains the concept. Use case-specific content that explains the concept. Use case-specific content that explains the concept."
-        />
-      </InterstitialScreenView>
+      <InterstitialScreen.Header
+        headerTitle={'headerTitle'}
+        headerSubTitle={'headerSubTitle'}
+      ></InterstitialScreen.Header>
+      <InterstitialScreen.Body
+        contentRenderer={(internalConfig) => (
+          <>
+            <InterstitialScreenView
+              stepTitle="Step 1"
+              translateWithId={translateWithId}
+            >
+              <InterstitialScreenViewModule
+                title={InterstitialScreenViewModuleTitle}
+                description="Use case-specific content that explains the concept. Use case-specific content that explains the concept. Use case-specific content that explains the concept. Use case-specific content that explains the concept. Use case-specific content that explains the concept."
+              />
+            </InterstitialScreenView>
+            <InterstitialScreenView
+              stepTitle="Step 2"
+              translateWithId={translateWithId}
+            >
+              <InterstitialScreenViewModule
+                title="Use case-specific heading 2"
+                description="Use case-specific content that explains the concept. Use case-specific content that explains the concept. Use case-specific content that explains the concept. Use case-specific content that explains the concept."
+              />
+            </InterstitialScreenView>
+          </>
+        )}
+      />
+      <InterstitialScreen.Footer />
     </InterstitialScreen>
   );
+};
 
 describe(componentName, () => {
   it('renders a component InterstitialScreen (Modal)', () => {
@@ -185,6 +213,7 @@ describe(componentName, () => {
       className: blockClass,
       interstitialAriaLabel: 'Modal Interstitial Screen',
     });
+
     expect(screen.getByText('Next'));
     expect(screen.getByText('Step 1'));
     const step1 = screen.getByText('Step 1');
@@ -194,18 +223,27 @@ describe(componentName, () => {
     expect(listElement1).toHaveClass(
       `${carbon.prefix}--progress-step--current`
     );
+    expect(listElement1).toHaveTextContent('Actuel');
+
     expect(listElement2).toHaveClass(
       `${carbon.prefix}--progress-step--incomplete`
     );
+    expect(listElement2).toHaveTextContent('Partiel');
+
     const nextButtonElement = screen.getByText('Next');
     expect(nextButtonElement).toHaveClass(`${blockClass}--next-btn`);
-    await act(() => userEvent.click(nextButtonElement));
+    await waitFor(() => userEvent.click(nextButtonElement, { timeout: 10 }));
+
     expect(listElement1).toHaveClass(
       `${carbon.prefix}--progress-step--complete`
     );
+    expect(listElement1).toHaveTextContent('Terminé');
+
     expect(listElement2).toHaveClass(
       `${carbon.prefix}--progress-step--current`
     );
+    expect(listElement2).toHaveTextContent('Actuel');
+
     expect(screen.getByText('Back')).toBeInTheDocument();
     const backButtonElement = screen.getByText('Back');
     expect(backButtonElement).toHaveClass(`${blockClass}--prev-btn`);
@@ -226,14 +264,5 @@ describe(componentName, () => {
     const closeBtn = screen.getByLabelText('Close');
     await act(() => userEvent.click(closeBtn));
     expect(onClose).toBeCalled();
-  });
-
-  it('renders an image with renderMedia prop', async () => {
-    renderComponent({
-      className: blockClass,
-      interstitialAriaLabel: 'Modal Interstitial Screen',
-      renderMedia: () => <img alt="img" />,
-    });
-    expect(screen.getByRole('img')).toBeInTheDocument();
   });
 });