+++ packages/ibm-products/src/components/OptionsTile/OptionsTile.tsx
@@ -13,7 +13,7 @@ import {
   WarningAltFilled,
   WarningFilled,
 } from '@carbon/react/icons';
-import { Layer, Toggle } from '@carbon/react';
+import { Heading, Layer, Section, Toggle } from '@carbon/react';
 import React, { MouseEvent, ReactNode, useRef, useState } from 'react';
 import { CarbonIconType } from '@carbon/icons-react/lib/CarbonIcon';
 import PropTypes from 'prop-types';
@@ -275,9 +275,9 @@ export let OptionsTile = React.forwardRef<HTMLDivElement, OptionsTileProps>(
 
       return (
         <div className={`${blockClass}__heading`}>
-          <h6 id={titleId} className={`${blockClass}__title`}>
+          <Heading id={titleId} className={`${blockClass}__title`}>
             {title}
-          </h6>
+          </Heading>
           {text && (
             <span className={summaryClasses} aria-hidden={summaryHidden}>
               {Icon && <Icon size={16} />}
@@ -289,7 +289,7 @@ export let OptionsTile = React.forwardRef<HTMLDivElement, OptionsTileProps>(
     };
 
     return (
-      <div
+      <Section
         {...rest}
         className={cx(blockClass, className, `${blockClass}--${size}`, {
           [`${blockClass}--closing`]: closing,
@@ -342,7 +342,7 @@ export let OptionsTile = React.forwardRef<HTMLDivElement, OptionsTileProps>(
         ) : (
           <div className={`${blockClass}__static-content`}>{renderTitle()}</div>
         )}
-      </div>
+      </Section>
     );
   }
 );