+++ packages/ibm-products/src/components/ProductiveCard/ProductiveCard.tsx
@@ -24,16 +24,16 @@ const componentName = 'ProductiveCard';
 type ActionIcon = {
   id?: string;
   icon?: CarbonIconType;
-  onKeyDown?(): void;
-  onClick?(): void;
+  onClick?: (event: MouseEvent) => void;
+  onKeydown?: (event: KeyboardEvent) => void;
   iconDescription?: string;
   href?: string;
 };
 type overflowAction = {
   id?: string;
   itemText?: string;
-  onClick?: () => void;
-  onKeydown?: () => void;
+  onClick?: (event: MouseEvent) => void;
+  onKeydown?: (event: KeyboardEvent) => void;
 };
 type PlacementType = 'top' | 'bottom';
 type ClickZoneType = 'one' | 'two' | 'three';
@@ -52,16 +52,9 @@ export interface ProductiveCardProps extends PropsWithChildren {
    */
   actionsPlacement?: PlacementType;
   /**
-   * Content that shows in the body of the card
-   */
-  // children: PropTypes.node,
-  /**
    * Optional user provided class
    */
   className?: string;
-
-  children: ReactNode;
-
   /**
    * Designates which zones of the card are clickable. Refer to design documentation for implementation guidelines
    */
@@ -78,7 +71,13 @@ export interface ProductiveCardProps extends PropsWithChildren {
   /**
    * Provides the callback for a clickable card
    */
-  onClick?: () => void;
+  onClick?: (event: MouseEvent) => void;
+
+  /**
+   * Provides the callback for keydown events on the card
+   */
+  onKeyDown?: (event: KeyboardEvent) => void;
+
   /**
    * Function that's called from the primary button or action icon
    */