+++ packages/ibm-products/src/custom-typings/index.d.ts
@@ -5,11 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {
-  AnimatePresenceProps,
-  MotionProps as OriginalMotionProps,
-  AnimatePresenceProps as PresenceProps,
-} from 'framer-motion';
 import React from 'react';
 
 declare module '@carbon/react' {
@@ -36,7 +31,6 @@ declare module '@carbon/react' {
     ComposedModal,
     ContainedList,
     ContainedListItem,
-    ContainedListItem,
     Content,
     ContentSwitcher,
     ContextMenu,
@@ -45,8 +39,6 @@ declare module '@carbon/react' {
     CopyButtonProps,
     DangerButton,
     DismissibleTag,
-    DataTable,
-    DataTable,
     DataTable,
     DataTableCustomRowData,
     DataTableCustomRowProps,
@@ -84,7 +76,6 @@ declare module '@carbon/react' {
     HeaderGlobalAction,
     HeaderGlobalBar,
     HeaderName,
-    HeaderName,
     HeaderNavigation,
     Heading,
     IconButton,
@@ -93,7 +84,6 @@ declare module '@carbon/react' {
     IdPrefix,
     InlineLoading,
     InlineNotification,
-    Layer,
     Layer,
     Link,
     LinkProps,
@@ -128,8 +118,6 @@ declare module '@carbon/react' {
     PopoverContent,
     PrimaryButton,
     ProgressBar,
-    ProgressBar,
-    ProgressIndicator,
     ProgressIndicator,
     ProgressIndicatorSkeleton,
     ProgressStep,
@@ -140,6 +128,7 @@ declare module '@carbon/react' {
     Search,
     SearchSkeleton,
     SecondaryButton,
+    Section,
     Select,
     SelectItem,
     SelectItemGroup,
@@ -160,28 +149,21 @@ declare module '@carbon/react' {
     StructuredListSkeleton,
     Switch,
     Tab,
-    Tab,
     TabContent,
     TabList,
     TabPanel,
     TabPanels,
     Table,
-    Table,
     TableBatchAction,
     TableBatchActions,
     TableBody,
-    TableBody,
-    TableCell,
     TableCell,
     TableContainer,
     TableExpandHeader,
     TableExpandRow,
     TableExpandedRow,
     TableHead,
-    TableHead,
     TableHeader,
-    TableHeader,
-    TableRow,
     TableRow,
     TableSelectAll,
     TableSelectAllProps,
@@ -189,12 +171,10 @@ declare module '@carbon/react' {
     TableToolbar,
     TableToolbarSearch,
     Tabs,
-    Tabs,
     TabsSkeleton,
     Tag,
     TextArea,
     TextAreaProps,
-    TextInput,
     TextInput,
     TextInputProps,
     Theme,
@@ -222,13 +202,3 @@ declare module '@carbon/react' {
 declare module '@carbon/colors';
 declare module '@carbon/motion';
 declare module '@carbon/feature-flags';
-
-declare module 'framer-motion' {
-  interface MotionProps extends OriginalMotionProps {
-    className?: string;
-    children?: React.ReactNode;
-  }
-  interface AnimatePresenceProps extends PresenceProps {
-    mode?: string;
-  }
-}